requirements.txt
pytest==7.1.2
requests==2.28.1
selenium==4.4.0
webdriver-manager==3.8.3
faker==13.15.1
allure-pytest==2.9.45